% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.r
\name{dispersal}
\alias{dispersal}
\alias{dispersal_linear}
\alias{dispersal_negexp}
\title{Generate dispersal kernel functions}
\usage{
dispersal_linear(dmax)

dispersal_negexp(alpha)
}
\arguments{
\item{dmax}{numeric; maximum dispersal distance.}

\item{alpha}{numeric; exponential decay rate parameter, equal to the inverse
of the mean dispersal distance.}
}
\value{
A dispersal kernel function.
}
\description{
Generate linear or negative exponential dispersal kernel functions, taking a
distance, \code{d}, and returning the probability density of dispersing that
distance.
}
\details{
The available dispersal kernel functions are:
\itemize{
\item \code{dispersal_linear}: decays linearly from 1 and \code{d = 0} to 0 at \code{d = dmax}.
\item \code{dispersal_negexp}: decays exponentially with rate parameter \code{alpha}.
}
}
\examples{
dl <- dispersal_linear(10)
dne <- dispersal_negexp(1/5)
dl(5)
dne(5)
}
