% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-capacity.r
\name{meta_capacity}
\alias{meta_capacity}
\title{Metapopulation capacity}
\usage{
meta_capacity(x, a, f, ex = 0.5, self = TRUE, patch_mc = FALSE, ...)
}
\arguments{
\item{x}{interpatch distance matrix or \link{patch_config} object.}

\item{a}{numeric; patch areas in the same order as the distance matrix.
Only used if `x`` is a distance matrix.}

\item{f}{function; dispersal kernel function taking a distance, \code{d}, and
returning the probability density of dispersing that distance.}

\item{ex}{numeric; extinction-area parameter, where extinction rate is given
by \eqn{E / A^ex} for patch size \eqn{A}.}

\item{self}{logical; whether patches can recolonize themselves, this
correction was proposed by Schnell et al. (2013) because the original
favours networks of small patches over large contiguous patches}

\item{patch_mc}{logical; whether to return the patch-level contributions to
the metapopulation capacity.}

\item{...}{additional parameters passed to the dispersal survival function.}
}
\value{
The metapopulation capacity or, if \code{patch_mc = TRUE}, a list
containing the metapopulation capacity (\code{capacity}) and the patch
contributions (\code{patch_mc}).
}
\description{
Metapopulation capacity (Hanski and Ovaskainen 2000) is a relative measure of
the ability of a spatially explicit landscape to support a metapopulation of
a species. This function implements the modification proposed by Schnell et
al. (2013) that includes self colonization.
}
\examples{
d <- matrix(c(0, 2, 4, 2, 0, 3, 4, 3, 0), nrow = 3)
a <- c(1, 1.5, 0.5)
f <- dispersal_negexp(1)
meta_capacity(d, a, f)

r <- raster::raster(nrows = 10, ncols = 10, crs = "+proj=laea")
r[] <- round(runif(raster::ncell(r)) * 0.7)
pc <- patch_config(r, "m")
f <- dispersal_negexp(1 / 100)
meta_capacity(pc, f = f)
}
\references{
\itemize{
\item Hanski, I, & O Ovaskainen (2000). The metapopulation capacity of a
fragmented landscape. Nature, 404(6779), 755.
\href{http://doi.org/10.1038/35008063}{doi:10.1038/35008063}
\item Schnell, JK, GM Harris, SL Pimm, & GJ Russell (2013). Estimating
Extinction Risk with Metapopulation Models of Large‐Scale Fragmentation.
Conservation Biology, 27(3), 520-530.
\href{http://doi.org/10.1111/cobi.12047}{doi:10.1111/cobi.12047}
}
}
