% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection-units.r
\name{projection_units}
\alias{projection_units}
\title{Get projection units of an sf object}
\usage{
projection_units(x)
}
\arguments{
\item{x}{an sf object}
}
\value{
A units object with the units of the projection.
}
\description{
Get projection units of an sf object
}
\examples{
pts <- sf::st_sfc(sf::st_point(c(0, 0)), sf::st_point(c(1, 1)))
pts_df <- sf::st_sf(a = 1:2, geom = pts)
sf::st_crs(pts_df) = "+proj=aea"
projection_units(pts_df)
}
