% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve-perimeter.r
\name{reserve_perimeter}
\alias{reserve_perimeter}
\title{Calculate the perimeter of a reserve network}
\usage{
reserve_perimeter(pu, x, units = c("km", "m"))
}
\arguments{
\item{pu}{\link[raster:RasterStack-class]{raster::RasterStack}, \link[sp:SpatialPolygonsDataFrame]{sp::SpatialPolygonsDataFrame}, or
\link[sf:sf]{sf::sf} object; planning units and representation levels of features as
layers (raster inputs) or columns (vector inputs).}

\item{x}{logical or binary; indicator variables specifying which planning
units are included.}

\item{units}{character; metapopulation capacity depends on the units used for
the areas and distances, this argument determines whether these are
measured in meters or kilometers.}
}
\value{
A \code{units} object giving the reserve perimeter in the specified units.
}
\description{
Calculate the perimeter of a reserve network
}
\examples{
r <- raster::raster(nrows = 10, ncols = 10, crs = "+proj=aea",
                    vals = sample(0:1, 100, replace = TRUE))
selected <- sample(c(FALSE, TRUE), 100, replace = TRUE, prob = c(0.7, 0.3))
reserve_perimeter(r, selected)
}
